Compatibility - update Galacticraft for oxygen compatibility with latest Mekanism versions. Fixes #1509

+++ dependencies/main/java/mekanism/api/EnumColor.java
@@ -92,7 +92,7 @@ public enum EnumColor
 	 */
 	public float getColor(int index)
 	{
-		return (float)rgbCode[index]/255F;
+		return rgbCode[index]/255F;
 	}
 
 	/**
